1.完成跳转到商品详情页，并实现参数的传递 2.修改首页夫分类，只显示推荐和上架

+++ user_service/src/main/java/com/stedu/mall/userservice/service/UserServiceImpl.java
@@ -44,7 +44,12 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public boolean delete(Integer id) {
+    public boolean delete(Integer id) throws SteduException {
+        //先注销再删除
+        User user = userMapper.selectById(id);
+        if (user.getStatus() != 3) {
+            throw new SteduException("请先注销后再删除");
+        }
         return userMapper.delete(id) == 1;
     }
